Translation update (bug fixes)

+++ src/main/java/dev/starless/maggiordomo/commands/interaction/filter/FilterInteraction.java
@@ -20,6 +20,7 @@ import net.dv8tion.jda.api.interactions.components.text.TextInputStyle;
 import net.dv8tion.jda.api.interactions.modals.Modal;
 import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder;
 import net.dv8tion.jda.api.utils.messages.MessageCreateData;
+import net.dv8tion.jda.api.utils.messages.MessageEditData;
 
 import java.util.Set;
 
@@ -34,11 +35,11 @@ public abstract class FilterInteraction implements Interaction {
         if (page != -1) {
             e.reply(createMenu(settings, page)).queue();
         } else {
-            TextInputStyle textStyle = type.equals(FilterType.CONTAINS) ? TextInputStyle.PARAGRAPH : TextInputStyle.SHORT;
-
+            TextInputStyle textStyle = type.equals(FilterType.BASIC) ? TextInputStyle.PARAGRAPH : TextInputStyle.SHORT;
+            Messages modalValueMessage = type.equals(FilterType.BASIC) ? Messages.COMMAND_FILTERS_BASIC_INPUT : Messages.COMMAND_FILTERS_PATTERN_INPUT;
             e.replyModal(Modal.create(getName(), Translations.string(Messages.FILTER_MENU_TITLE, settings.getLanguage()))
                             .addActionRow(TextInput.create("input", "input", textStyle)
-                                    .setValue(Translations.string(Messages.FILTER_EXPLANATION, settings.getLanguage()))
+                                    .setValue(Translations.string(modalValueMessage, settings.getLanguage()))
                                     .setRequiredRange(1, 256)
                                     .build())
                             .build())
@@ -52,8 +53,14 @@ public abstract class FilterInteraction implements Interaction {
     public VC onModalInteraction(VC vc, Settings settings, String id, ModalInteractionEvent e) {
         onInputReceived(settings, e);
 
-        e.getMessage().delete().queue();
-        e.reply(createMenu(settings, 0)).queue();
+        if (e.isAcknowledged()) {
+            e.getMessage().editMessage(MessageEditData.fromCreateData(createMenu(settings, 0)))
+                    .setReplace(true)
+                    .queue();
+        } else {
+            e.getMessage().delete().queue();
+            e.reply(createMenu(settings, 0)).queue();
+        }
 
         return null;
     }
@@ -87,8 +94,8 @@ public abstract class FilterInteraction implements Interaction {
         Set<String> words = settings.getFilterWords(type);
         int maxPages = (int) Math.ceil(words.size() / 10D);
 
-        String filterName = Translations.string(type.equals(FilterType.CONTAINS) ? Messages.FILTER_BASIC : Messages.FILTER_PATTERN, settings.getLanguage());
-        String content = Translations.string(Messages.COMMAND_SETUP_EXPLANATION, settings.getLanguage(), filterName);
+        String filterName = Translations.string(type.equals(FilterType.BASIC) ? Messages.FILTER_BASIC : Messages.FILTER_PATTERN, settings.getLanguage());
+        String content = Translations.string(Messages.COMMAND_FILTERS_EXPLANATION, settings.getLanguage(), filterName);
 
         MessageCreateBuilder builder = new MessageCreateBuilder().setContent(content);
